requirements.txt
certifi
chardet
configparser
Deprecated
idna
PyGithub
PyJWT
requests
urllib3
wrapt
gitpython
